Renamed auth-page to prevent confusion with 'userAccount' page in organization. #9
