+++ client/pages/auth-page.jsx
